requirements.txt
agate==1.9.1
alembic==1.9.2
annotated-types==0.7.0
anyio==3.6.2
apache-airflow==2.5.1
apache-airflow-providers-common-sql==1.3.3
apache-airflow-providers-ftp==3.3.0
apache-airflow-providers-http==4.1.1
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.5
cachelib==0.9.0
cattrs==22.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron_descriptor==1.2.32
croniter==1.3.8
cryptography==42.0.8
daff==1.3.46
dbt==1.0.0.38.4
dbt-adapters==1.3.2
dbt-common==1.4.0
dbt-core==1.8.3
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
dbt-snowflake==1.8.3
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fqdn==1.5.1
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keyring==24.3.1
lazy-object-proxy==1.9.0
leather==0.4.0
linkify-it-py==2.0.0
lockfile==0.12.2
Logbook==1.5.3
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==3.13.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mistune==3.0.2
more-itertools==10.3.0
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.1
notebook_shim==0.2.4
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
pickleshare==0.7.5
 
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.0.0
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.14.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==2.3.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.1.0
rpds-py==0.18.1
SecretStorage==3.3.3
Send2Trash==1.8.3
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.39.0
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.2.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.12.5
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==1.26.19
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.11
wrapt==1.14.1
WTForms==3.0.1
zipp==3.19.2
